requirements.txt
click==8.1.3
emoji==1.7.0
prometheus-client==0.14.1
PyGithub==1.55
python-telegram-bot==13.11
steam[client]==1.2.0
